Create code for the app in android studio
This will be the code for the Main Activity.  this will be the content of MainActivity.java. For this app there will be two Activity and each activity will have different layout. We use explicit intent for travelling from one Activity to the other. For moving from the application to the outside application for additional purpose , we use implicit intent.